Create module gpu_bridge
