+++ crates/re_viewer/src/gpu_bridge/tensor_to_gpu.rs
@@ -1,3 +1,5 @@
+//! Upload [`Tensor`] to [`re_renderer`].
+
 use std::borrow::Cow;
 
 use bytemuck::{allocation::pod_collect_to_vec, cast_slice, Pod};
@@ -7,11 +9,13 @@ use wgpu::TextureFormat;
 use re_log_types::component_types::{Tensor, TensorData};
 use re_renderer::{
     renderer::{ColorMapper, ColormappedTexture},
-    resource_managers::{GpuTexture2DHandle, Texture2DCreationDesc},
+    resource_managers::Texture2DCreationDesc,
     RenderContext,
 };
 
-use super::caches::TensorStats;
+use crate::misc::caches::TensorStats;
+
+use super::get_or_create_texture;
 
 // ----------------------------------------------------------------------------
 
@@ -380,18 +384,6 @@ fn general_texture_creation_desc_from_tensor<'a>(
         width,
         height,
     })
-}
-
-pub fn get_or_create_texture<'a, Err>(
-    render_ctx: &mut RenderContext,
-    texture_key: u64,
-    try_create_texture_desc: impl FnOnce() -> Result<Texture2DCreationDesc<'a>, Err>,
-) -> Result<GpuTexture2DHandle, Err> {
-    render_ctx.texture_manager_2d.get_or_create_with(
-        texture_key,
-        &mut render_ctx.gpu_resources.textures,
-        try_create_texture_desc,
-    )
 }
 
 fn cast_slice_to_cow<From: Pod>(slice: &[From]) -> Cow<'_, [u8]> {
