+++ crates/re_viewer/src/ui/view_tensor/gpu.rs
@@ -72,7 +72,7 @@ fn upload_texture_slice_to_gpu(
 ) -> Result<re_renderer::resource_managers::GpuTexture2DHandle, TensorUploadError> {
     let id = egui::util::hash((tensor.id(), slice_selection));
 
-    crate::misc::tensor_to_gpu::get_or_create_texture(render_ctx, id, || {
+    crate::gpu_bridge::get_or_create_texture(render_ctx, id, || {
         texture_desc_from_tensor(tensor, slice_selection)
     })
 }
